kloud/koding: move queue into his own file
